% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_headcount_ratio.R
\name{compute_headcount_ratio}
\alias{compute_headcount_ratio}
\title{Compute headcount ratio}
\usage{
compute_headcount_ratio(.data, ..., censored = T)
}
\arguments{
\item{.data}{data}

\item{...}{others}

\item{censored}{boolean}
}
\description{
Compute headcount ratio
}
