% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deprivation_matrix.R
\name{create_deprivation_matrix}
\alias{create_deprivation_matrix}
\title{Create deprivation matrix}
\usage{
create_deprivation_matrix(
  .data,
  .deprivation_cutoffs,
  ...,
  .mpi_specs = getOption("mpi_specs")
)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{.deprivation_cutoffs}{A list of data frames or tibbles containing the deprivation cutoffs}

\item{...}{A tidyselect specification of columns to select}

\item{.mpi_specs}{A data frame or tibble containing the specifications of the MPI}
}
\value{
A list
}
\description{
Create deprivation matrix
}
