% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_deprivation_cutoff.R
\name{define_deprivation_cutoff}
\alias{define_deprivation_cutoff}
\title{Define deprivation cutoffs}
\usage{
define_deprivation_cutoff(
  .data,
  .indicator,
  .condition,
  .mpi_specs = getOption("mpi_specs"),
  .collapse = FALSE,
  .collapse_condition = NULL
)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{.indicator}{A string or symbol indicating the name of indicator to be used}

\item{.condition}{A string or symbol indicating the condition to be used}

\item{.mpi_specs}{A data frame or tibble containing specifications of the MPI}

\item{.collapse}{A boolean indicating whether to collapse the data frame or not}

\item{.collapse_condition}{A string or symbol indicating the condition to be used when collapsing the data}
}
\description{
Define deprivation cutoffs
}
