% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_sentence_case}
\alias{to_sentence_case}
\title{Convert to sentence case}
\usage{
to_sentence_case(.x, words_to_preserve = NULL)
}
\arguments{
\item{words_to_preserve}{}
}
\description{
Convert to sentence case
}
