% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-mpi-specs.R
\name{use_global_mpi_specs}
\alias{use_global_mpi_specs}
\title{Use Global MPI specification}
\usage{
use_global_mpi_specs(...)
}
\arguments{
\item{...}{Passed to \code{\link{global_mpi_specs}}.}
}
\value{
An \code{mpi_specs} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
  Please use \code{\link{global_mpi_specs}} instead.
}
\examples{
\dontrun{
# Deprecated — use global_mpi_specs() instead
mpi_specs <- use_global_mpi_specs(uid = "uuid")
}

}
