% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mpi.R
\name{compute_mpi}
\alias{compute_mpi}
\title{Compute MPI}
\usage{
compute_mpi(
  .deprivation_matrix,
  ...,
  .mpi_specs = getOption("mpi_specs"),
  .names_separator = ">",
  .generate_output = FALSE,
  .output_filename = NULL
)
}
\arguments{
\item{.deprivation_matrix}{Deprivation matrix}

\item{...}{Grouping columns}

\item{.mpi_specs}{A data frame or tibble containing specifications of the MPI}

\item{.names_separator}{Column separator that defines the hierarchy of the column header.}

\item{.generate_output}{Whether to generate an output}

\item{.output_filename}{Output filename}
}
\description{
Compute MPI
}
