% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mpi.R
\name{compute_mpi}
\alias{compute_mpi}
\title{Compute MPI}
\usage{
compute_mpi(
  .data,
  .deprivation_cutoffs,
  ...,
  .mpi_specs = getOption("mpi_specs"),
  .names_separator = ">",
  .include_deprivation_matrix = TRUE,
  .generate_output = FALSE,
  .formatted_output = TRUE,
  .output_filename = NULL
)
}
\arguments{
\item{.data}{A data frame}

\item{.deprivation_cutoffs}{list of deprivation cutoffs}

\item{...}{Grouping columns}

\item{.mpi_specs}{A data frame or tibble containing specifications of the MPI}

\item{.names_separator}{Column separator that defines the hierarchy of the column header.}

\item{.include_deprivation_matrix}{Include deprivation matrix}

\item{.generate_output}{Whether to generate an output}

\item{.formatted_output}{Whether formatting is to be applied to the output}

\item{.output_filename}{Output filename}
}
\description{
Compute MPI
}
