% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_mpi_specs.R
\name{define_mpi_specs}
\alias{define_mpi_specs}
\title{Define MPI specifications: dimensions, indicators, and weights}
\usage{
define_mpi_specs(
  .mpi_specs_file,
  .poverty_cutoffs = 1/3,
  .unit_of_analysis = "household",
  .id = NULL,
  .source_of_data = NULL,
  .names_separator = ">"
)
}
\arguments{
\item{.mpi_specs_file}{Accepts \code{.xlsx} (Excel), \code{.json}, \code{.csv}, or \code{.txt} (TSV) file format. This file should contain the following columns/variables: \code{Dimension}, \code{Indicator}, \code{Variable}, \code{Weight}, and \code{Description} (optional). Click \href{https://github.com/yng-me/mpi/blob/main/tests/data}{here} to see some examples.}

\item{.poverty_cutoffs}{Accepts single value or a vector of poverty cutoffs. See example below.}

\item{.unit_of_analysis}{Accepts  \code{individual} | \code{family} | \code{household} | \code{community}. Default value is \code{household}.}

\item{.id}{Column name containing unique ID of the observation which defines the lowest level of disaggregation.}

\item{.source_of_data}{Source of data used in the computation}

\item{.names_separator}{Column separator that defines the hierarchy of the column header.}
}
\description{
Use this function to define the identified MPI dimensions, indicators and its corresponding weights using any of the accessible file types: \code{.xlsx} (Excel), \code{.json}, \code{.csv}, or \code{.txt} (TSV). You can also set the poverty cutoff or list of poverty cutoffs (to achieve gradient list of MPIs) that will be used in the computation of MPI.
}
