% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-mpi-specs.R
\name{use_global_mpi_specs}
\alias{use_global_mpi_specs}
\title{Use Global MPI specification}
\usage{
use_global_mpi_specs(...)
}
\arguments{
\item{...}{Accepts all arguments in \code{define_mpi_specs}}
}
\value{
Global MPI specs
}
\description{
Use built-in specification file for Global MPI.
}
\examples{

use_global_mpi_specs()
}
